Separate main.js into some modules.
Helper functions and the plugin function are separated.
